Implement persist options and implement this option for background jobs

+++ src/nimbo/main.py
@@ -159,6 +159,8 @@ def delete_instance(instance_id, dry_run):
 @click.option("--dry-run", is_flag=True)
 def delete_all_instances(dry_run):
     """Terminates all your instances."""
+    click.confirm("This will delete all your running instances.\n"
+                  "Do you want to continue?", abort=True)
     session, config = get_session_and_config_minimal()
     utils.delete_all_instances(session, config, dry_run)
 
@@ -176,7 +178,8 @@ def create_bucket(bucket_name):
 
 @cli.command()
 @click.argument("folder", type=click.Choice(["datasets", "results"]), required=True)
-@click.option("--delete", is_flag=True)
+@click.option("--delete", is_flag=True,
+              help="Deletes any files that exist in the local folder but don't exist in the remote folder.")
 def push(folder, delete):
     """Push your local datasets/results folder onto S3."""
     session, config = get_session_and_config_storage()
@@ -185,10 +188,10 @@ def push(folder, delete):
 
 @cli.command()
 @click.argument("folder", type=click.Choice(["datasets", "results"]), required=True)
-@click.option("--delete", is_flag=True)
+@click.option("--delete", is_flag=True, 
+              help="Deletes any files that exist in the local folder but don't exist in the remote folder.")
 def pull(folder, delete):
-    """Pull the S3 datasets/results folder into your local computer.
-    """
+    """Pull the S3 datasets/results folder into your local computer."""
     session, config = get_session_and_config_storage()
     storage.pull(session, config, folder, delete)
